requirements.txt
numpy
Pillow
PyYAML
einops
huggingface_hub
timm
torch
